pull code a bit apart
